refactor: 기본정보, 경력/지원/기술 정보 component 분리
