+++ src/components/domain/portfolio/BasicInfoSection.module.css
@@ -1,0 +1,18 @@
+.sec__profile {
+  display: flex;
+  align-items: center;
+  gap: 1rem;
+  margin-bottom: var(--sp-8);
+
+  img {
+    width: 6.25rem;
+    height: 6.25rem;
+    border-radius: 50%;
+    background-color: #fff;
+    box-shadow: 0 4px 12px rgba(0, 0, 0, 0.15);
+  }
+}
+
+.sec__profile__info {
+  font-size: var(--fs-md);
+}
