+++ src/pages/Portfolio/NewPortfolio.tsx
@@ -1,12 +1,9 @@
 import Button from '@/components/common/Button'
 import S from './NewPortfolio.module.css'
 import Input from '@/components/common/Input'
-import RadioGroup from '@/components/common/RadioGroup'
 import { useCallback, useEffect, useRef, useState } from 'react'
-import CheckboxSelect from '@/components/common/CheckboxSelect'
 import Textarea from '@/components/common/Textarea'
 import FileUploader from '@/components/common/FileUploader'
-import ImageUploader from '@/components/common/ImageUploader'
 import type { PortfolioData, TempItem } from '@/types/portfolio'
 import { useSavePortfolio } from '@/hooks/portfolio/useSavePortfolio'
 import { useCheckValidation } from '@/hooks/portfolio/useCheckValidation'
@@ -17,6 +14,8 @@ import { useStickyMenu } from '@/hooks/portfolio/useStickyMenu'
 import SideTempList from './SideTempList'
 import supabase from '@/lib/supabaseClient'
 import { useSearchParams } from 'react-router-dom'
+import TechProfileSection from '@/components/domain/portfolio/TechInfoSection'
+import BasicInfoSection from '@/components/domain/portfolio/BasicInfoSection'
 
 // TODOS : 기본정보 마이페이지에서 불러와야함
 const TempData: PortfolioData = {
@@ -39,28 +38,6 @@ const TempData: PortfolioData = {
 }
 
 export const NewPortfolio = () => {
-  /* --- 경력 수준 라디오 그룹 상태 및 옵션 --- */
-  const CAREER_SELECT = [
-    { label: '신입 (0년)', value: 'zeroYear' },
-    { label: '1년차 이상', value: 'oneYearsPlus' },
-    { label: '3년차 이상', value: 'threeYearsPlus' },
-    { label: '5년차 이상', value: 'fiveYearsPlus' },
-    { label: '10년차 이상', value: 'tenYearsPlus' }
-  ]
-
-  /* --- 지원분야 라디오 그룹 상태 및 옵션 --- */
-  const INTEREST_SELECT = [
-    { label: '프론트엔드 개발', value: 'FE' },
-    { label: '백엔드 개발', value: 'BE' },
-    { label: '풀스택 개발', value: 'FullStack' },
-    { label: '모바일 개발', value: 'Mobile' },
-    { label: '임베디드 개발', value: 'Embedded' },
-    { label: 'UI/UX 디자인', value: 'UIUX' },
-    { label: '그래픽 디자인', value: 'Graphic' },
-    { label: '모션 디자인', value: 'Motion' },
-    { label: '일러스트', value: 'Illustration' }
-  ]
-
   const [portfolioData, setPortfolioData] = useState<PortfolioData>(TempData)
 
   /* --- 마이페이지 임시저장글 불러오기 --- */
@@ -115,7 +92,7 @@ export const NewPortfolio = () => {
   }, [userInfo])
 
   /* --- 입력값 변경 시 상태 저장 및 관련 에러 제거 --- */
-  const { handleChangeForm, handelChangeRadio } = usePortfolioForm(
+  const { handleChangeForm, handleChangeRadio } = usePortfolioForm(
     setPortfolioData,
     setErrors
   )
@@ -167,7 +144,7 @@ export const NewPortfolio = () => {
 
       if (error) throw error
       if (!data) return
-      console.log(typeof data.techStack, data.techStack)
+
       setErrors({})
 
       setPortfolioData(prev => ({
@@ -274,76 +251,17 @@ export const NewPortfolio = () => {
         </div>
       </div>
       <form>
-        <section className={S['sec']}>
-          <h3 className="a11y-hidden">기본정보</h3>
-          <div className={S['sec__profile']}>
-            <ImageUploader
-              key={portfolioData.id}
-              id="exImage"
-              value={portfolioData.profileImage}
-              onChange={image => handleChangeForm('profileImage', image)}
-            />
-            <dl className={S['sec__profile__info']}>
-              <dt className="a11y-hidden">이름</dt>
-              <dd>{portfolioData.name}</dd>
+        <BasicInfoSection
+          portfolioData={portfolioData}
+          handleChangeForm={handleChangeForm}
+        />
 
-              <dt className="a11y-hidden">생년월일</dt>
-              <dd>{portfolioData.birthDate}</dd>
-            </dl>
-          </div>
-
-          <div className={S.sec__form}>
-            <Input
-              id="exId01"
-              label="이메일"
-              value={portfolioData.email}
-              readOnly
-            />
-
-            <Input
-              type="tel"
-              id="exId02"
-              label="전화번호"
-              value={portfolioData.phone}
-              readOnly
-            />
-          </div>
-        </section>
-
-        <section className={S['sec']}>
-          <h3 className="a11y-hidden">지원분야 / 경력 수준 / 기술 스택</h3>
-          <div className={S['sec__form']}>
-            <RadioGroup
-              label="경력수준"
-              name="careerOption"
-              options={CAREER_SELECT}
-              checked={portfolioData.career.value}
-              onChange={e =>
-                handelChangeRadio('career', e.target.value, CAREER_SELECT)
-              }
-              error={errors.career}
-            />
-
-            <RadioGroup
-              label="지원분야"
-              name="fieldOfSupport"
-              options={INTEREST_SELECT}
-              checked={portfolioData.interest.value}
-              onChange={e =>
-                handelChangeRadio('interest', e.target.value, INTEREST_SELECT)
-              }
-              error={errors.interest}
-            />
-
-            <CheckboxSelect
-              value={portfolioData.techStack}
-              onChange={stack => {
-                handleChangeForm('techStack', stack)
-              }}
-              error={errors.techStack}
-            />
-          </div>
-        </section>
+        <TechProfileSection
+          portfolioData={portfolioData}
+          handleChangeRadio={handleChangeRadio}
+          handleChangeForm={handleChangeForm}
+          errors={errors ?? {}}
+        />
 
         <section className={S['sec']}>
           <h3 className="a11y-hidden">포트폴리오 소개</h3>
